+++ packages/guide/CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added new type of generic cooking ingredients
 - Added all known festival shops to locations
 - Added Raj's Coffee Corner and White Flamingo to shops
+- Added settings to delete checklists
 
 #### Changed